requirements.txt
streamlit
groq
langchain
fastapi
uvicorn
beautifulsoup4
requests
python-dotenv